+++ src/entities/ImgBufferOpenGL_t.cpp
@@ -5,9 +5,9 @@
 using APTracer::Entities::ImgBufferOpenGL_t;
 using APTracer::Entities::Vec3f;
 
-ImgBufferOpenGL_t::ImgBufferOpenGL_t(unsigned int size_x, unsigned int size_y): ImgBuffer_t(size_x, size_y) {
+ImgBufferOpenGL_t::ImgBufferOpenGL_t(size_t size_x, size_t size_y): ImgBuffer_t(size_x, size_y) {
     img_gl_ = new unsigned char[size_x_*size_y_*3];
-    for (unsigned int i = 0; i < size_x_*size_y_*3; ++i) {
+    for (size_t i = 0; i < size_x_*size_y_*3; ++i) {
         img_gl_[i] = 0;
     }
     tex_ = 0;
@@ -19,15 +19,15 @@ ImgBufferOpenGL_t::~ImgBufferOpenGL_t() {
 
 void ImgBufferOpenGL_t::reset() {
     updates_ = 0;
-    for (unsigned int j = 0; j < size_y_*size_x_; ++j) {
+    for (size_t j = 0; j < size_y_*size_x_; ++j) {
         img_[j] = Vec3f();
     }
 }
 
-void ImgBufferOpenGL_t::update(const Vec3f* img, unsigned int size_x, unsigned int size_y) {
+void ImgBufferOpenGL_t::update(const Vec3f* img, size_t size_x, size_t size_y) {
     ++updates_;
 
-    for (unsigned int j = 0; j < size_y*size_x; ++j) {
+    for (size_t j = 0; j < size_y*size_x; ++j) {
         img_[j] += img[j];
     }
 }
@@ -70,7 +70,7 @@ void ImgBufferOpenGL_t::update() {
     glutSwapBuffers();
 }
 
-void ImgBufferOpenGL_t::update(const Vec3f &colour, unsigned int pos_x, unsigned int pos_y) {
+void ImgBufferOpenGL_t::update(const Vec3f &colour, size_t pos_x, size_t pos_y) {
     //img_[pos_y][pos_x] = img_[pos_y][pos_x] * (1.0 - 1.0/static_cast<double>(updates_)) + colour/static_cast<double>(updates_);
     img_[pos_y*size_x_ + pos_x] += colour;
     img_gl_[((size_y_-1-pos_y)*size_x_ + pos_x)*3] = static_cast<unsigned char>(std::lround(std::max(std::min(1.0, img_[pos_y*size_x_ + pos_x][0]/updates_), 0.0) * 255.0));
@@ -78,14 +78,14 @@ void ImgBufferOpenGL_t::update(const Vec3f &colour, unsigned int pos_x, unsigned
     img_gl_[((size_y_-1-pos_y)*size_x_ + pos_x)*3+2] = static_cast<unsigned char>(std::lround(std::max(std::min(1.0, img_[pos_y*size_x_ + pos_x][2]/updates_), 0.0) * 255.0));
 }
 
-void ImgBufferOpenGL_t::set(const Vec3f* img, unsigned int size_x, unsigned int size_y) {
+void ImgBufferOpenGL_t::set(const Vec3f* img, size_t size_x, size_t size_y) {
     updates_ = 1;
-    for (unsigned int j = 0; j < size_y*size_x; ++j) {
+    for (size_t j = 0; j < size_y*size_x; ++j) {
         img_[j] = img[j];
     }
 }
 
-void ImgBufferOpenGL_t::set(const Vec3f &colour, unsigned int pos_x, unsigned int pos_y) {
+void ImgBufferOpenGL_t::set(const Vec3f &colour, size_t pos_x, size_t pos_y) {
     img_[pos_y*size_x_ + pos_x] = colour;
     img_gl_[((size_y_-1-pos_y)*size_x_ + pos_x)*3] = static_cast<unsigned char>(std::lround(std::max(std::min(1.0, img_[pos_y*size_x_ + pos_x][0]/updates_), 0.0) * 255.0));
     img_gl_[((size_y_-1-pos_y)*size_x_ + pos_x)*3+1] = static_cast<unsigned char>(std::lround(std::max(std::min(1.0, img_[pos_y*size_x_ + pos_x][1]/updates_), 0.0) * 255.0));
